requirements.txt
streamlit>=1.40.0
openai>=1.50.0
mcp[cli]>=1.0.0
httpx>=0.27.0
pydantic>=2.0.0
python-dotenv>=1.0.0
